Fit a Model Specification to a Dataset

fit() and fit_xy() take a model specification, translate_celery the required code by substituting arguments, and execute the model fit routine.

Usage

# S3 method for cluster_spec
fit(object, formula, data, control = control_celery(), ...)

# S3 method for cluster_spec
fit_xy(object, x, control = control_celery(), ...)

Arguments

object

An object of class cluster_spec that has a chosen engine (via set_engine_celery()).

formula

An object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data

Optional, depending on the interface (see Details below). A data frame containing all relevant variables (e.g. outcome(s), predictors, case weights, etc). Note: when needed, a named argument should be used.

control

A named list with elements verbosity and catch. See control_celery().

...

Not currently used; values passed here will be ignored. Other options required to fit the model should be passed using set_engine_celery().

x

A matrix, sparse matrix, or data frame of predictors. Only some models have support for sparse matrix input. See celery::get_encoding_celery() for details. x should have column names.

Value

A cluster_fit object that contains several elements:

lvl: If the outcome is a factor, this contains the factor levels at the time of model fitting.

spec: The model specification object (object in the call to fit)

fit: when the model is executed without error, this is the model object. Otherwise, it is a try-error object with the error message.

preproc: any objects needed to convert between a formula and non-formula interface (such as the terms object)

The return value will also have a class related to the fitted model (e.g. "_glm") before the base class of "cluster_fit".

Details

fit() and fit_xy() substitute the current arguments in the model specification into the computational engine's code, check them for validity, then fit the model using the data and the engine-specific code. Different model functions have different interfaces (e.g. formula or x/y) and these functions translate_celery between the interface used when fit() or fit_xy() was invoked and the one required by the underlying model.

When possible, these functions attempt to avoid making copies of the data. For example, if the underlying model uses a formula and fit() is invoked, the original data are references when the model is fit. However, if the underlying model uses something else, such as x/y, the formula is evaluated and the data are converted to the required format. In this case, any calls in the resulting model objects reference the temporary objects used to fit the model.

If the model engine has not been set, the model's default engine will be used (as discussed on each model page). If the verbosity option of control_celery() is greater than zero, a warning will be produced.

If you would like to use an alternative method for generating contrasts when supplying a formula to fit(), set the global option contrasts to your preferred method. For example, you might set it to: options(contrasts = c(unordered = "contr.helmert", ordered = "contr.poly")). See the help page for stats::contr.treatment() for more possible contrast types.

See also

Examples

# Although `glm()` only has a formula interface, different
# methods for specifying the model can be used

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

kmeans_mod <- k_means(k = 5)

using_formula <-
  kmeans_mod %>%
  set_engine_celery("stats") %>%
  fit(~., data = mtcars)

using_x <-
  kmeans_mod %>%
  set_engine_celery("stats") %>%
  fit_xy(x = mtcars)

using_formula
#> $spec
#> K Means Cluster Specification (partition)
#> 
#> Main Arguments:
#>   k = 5
#> 
#> Computational engine: stats 
#> 
#> Model fit template:
#> stats::kmeans(x = missing_arg(), centers = missing_arg(), centers = 5)
#> 
#> $fit
#> K-means clustering with 5 clusters of sizes 4, 7, 10, 4, 7
#> 
#> Cluster means:
#>        mpg      cyl     disp       hp     drat       wt     qsec        vs
#> 1 14.60000 8.000000 340.5000 272.2500 3.675000 3.537500 15.08750 0.0000000
#> 2 19.94286 5.714286 166.5714 120.1429 3.705714 3.107857 18.47143 0.5714286
#> 3 27.05000 4.000000 101.5700  81.4000 4.086000 2.199300 18.76100 0.9000000
#> 4 13.67500 8.000000 443.0000 206.2500 3.060000 4.966000 17.56750 0.0000000
#> 5 17.10000 7.714286 295.3429 160.7143 3.050000 3.598571 17.66143 0.1428571
#>          am gear     carb
#> 1 0.5000000  4.0 5.000000
#> 2 0.4285714  4.0 3.571429
#> 3 0.8000000  4.1 1.500000
#> 4 0.0000000  3.0 3.500000
#> 5 0.0000000  3.0 2.285714
#> 
#> Clustering vector:
#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>                   2                   2                   3                   5 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>                   5                   2                   1                   3 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>                   2                   2                   2                   5 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>                   5                   5                   4                   4 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>                   4                   3                   3                   3 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>                   3                   5                   5                   1 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>                   4                   3                   3                   3 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>                   1                   2                   1                   3 
#> 
#> Within cluster sum of squares by cluster:
#> [1]  7654.146  8808.032 10247.471  4665.041 11474.702
#>  (between_SS / total_SS =  93.1 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#> $elapsed
#> $elapsed$elapsed
#> [1] NA
#> 
#> 
#> $preproc
#> $preproc$offset
#> NULL
#> 
#> $preproc$terms
#> ~mpg + cyl + disp + hp + drat + wt + qsec + vs + am + gear + 
#>     carb
#> attr(,"variables")
#> list(mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb)
#> attr(,"factors")
#>      mpg cyl disp hp drat wt qsec vs am gear carb
#> mpg    1   0    0  0    0  0    0  0  0    0    0
#> cyl    0   1    0  0    0  0    0  0  0    0    0
#> disp   0   0    1  0    0  0    0  0  0    0    0
#> hp     0   0    0  1    0  0    0  0  0    0    0
#> drat   0   0    0  0    1  0    0  0  0    0    0
#> wt     0   0    0  0    0  1    0  0  0    0    0
#> qsec   0   0    0  0    0  0    1  0  0    0    0
#> vs     0   0    0  0    0  0    0  1  0    0    0
#> am     0   0    0  0    0  0    0  0  1    0    0
#> gear   0   0    0  0    0  0    0  0  0    1    0
#> carb   0   0    0  0    0  0    0  0  0    0    1
#> attr(,"term.labels")
#>  [1] "mpg"  "cyl"  "disp" "hp"   "drat" "wt"   "qsec" "vs"   "am"   "gear"
#> [11] "carb"
#> attr(,"order")
#>  [1] 1 1 1 1 1 1 1 1 1 1 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 0
#> attr(,".Environment")
#> <environment: 0x561d6871f110>
#> attr(,"predvars")
#> list(mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb)
#> attr(,"dataClasses")
#>       mpg       cyl      disp        hp      drat        wt      qsec        vs 
#> "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" "numeric" 
#>        am      gear      carb 
#> "numeric" "numeric" "numeric" 
#> 
#> $preproc$xlevels
#> named list()
#> 
#> $preproc$options
#> $preproc$options$indicators
#> [1] "traditional"
#> 
#> $preproc$options$composition
#> [1] "matrix"
#> 
#> $preproc$options$remove_intercept
#> [1] TRUE
#> 
#> 
#> 
#> attr(,"class")
#> [1] "_kmeans"     "cluster_fit"
using_x
#> $spec
#> K Means Cluster Specification (partition)
#> 
#> Main Arguments:
#>   k = 5
#> 
#> Computational engine: stats 
#> 
#> Model fit template:
#> stats::kmeans(x = missing_arg(), centers = missing_arg(), centers = 5)
#> 
#> $fit
#> K-means clustering with 5 clusters of sizes 7, 10, 4, 4, 7
#> 
#> Cluster means:
#>        mpg      cyl     disp       hp     drat       wt     qsec        vs
#> 1 19.94286 5.714286 166.5714 120.1429 3.705714 3.107857 18.47143 0.5714286
#> 2 27.05000 4.000000 101.5700  81.4000 4.086000 2.199300 18.76100 0.9000000
#> 3 13.67500 8.000000 443.0000 206.2500 3.060000 4.966000 17.56750 0.0000000
#> 4 14.60000 8.000000 340.5000 272.2500 3.675000 3.537500 15.08750 0.0000000
#> 5 17.10000 7.714286 295.3429 160.7143 3.050000 3.598571 17.66143 0.1428571
#>          am gear     carb
#> 1 0.4285714  4.0 3.571429
#> 2 0.8000000  4.1 1.500000
#> 3 0.0000000  3.0 3.500000
#> 4 0.5000000  4.0 5.000000
#> 5 0.0000000  3.0 2.285714
#> 
#> Clustering vector:
#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>                   1                   1                   2                   5 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>                   5                   1                   4                   2 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>                   1                   1                   1                   5 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>                   5                   5                   3                   3 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>                   3                   2                   2                   2 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>                   2                   5                   5                   4 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>                   3                   2                   2                   2 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>                   4                   1                   4                   2 
#> 
#> Within cluster sum of squares by cluster:
#> [1]  8808.032 10247.471  4665.041  7654.146 11474.702
#>  (between_SS / total_SS =  93.1 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#> $elapsed
#> $elapsed$elapsed
#> [1] NA
#> 
#> 
#> attr(,"class")
#> [1] "_kmeans"     "cluster_fit"