Declare a computational engine and specific arguments

set_engine_celery() is used to specify which package or system will be used to fit the model, along with any arguments specific to that software.

Usage

set_engine_celery(object, engine, ...)

Arguments

object

A model specification.

engine

A character string for the software that should be used to fit the model. This is highly dependent on the type of model (e.g. linear regression, random forest, etc.).

...

Any optional arguments associated with the chosen computational engine. These are captured as quosures and can be tune().

Value

An updated model specification.

Engines

Based on the currently loaded packages, the following lists the set of engines available to each model specification.

Examples

# First, set general arguments using the standardized names
mod <-
  k_means(k = 10) %>%
  # now say how you want to fit the model and another other options
  set_engine_celery("stats", iter.max = 15)
translate_celery(mod, engine = "stats")
#> K Means Cluster Specification (partition)
#> 
#> Main Arguments:
#>   k = 10
#> 
#> Engine-Specific Arguments:
#>   iter.max = 15
#> 
#> Computational engine: stats 
#> 
#> Model fit template:
#> stats::kmeans(x = missing_arg(), centers = missing_arg(), centers = 10, 
#>     iter.max = 15)